requirements.txt
python-dotenv
openai
supabase
beautifulsoup4
requests
feedparser